+++ prPractica01/src/prPractica01/Ejercicioo14.java
@@ -14,21 +14,22 @@ public class Ejercicioo14 {
 
 		Scanner sc = new Scanner(System.in);
 		
-		int num1,num2;
+		float num1,num2;
 		
 		System.out.print("Introduce el primer número: ");
 		
-		num1=sc.nextInt();
+		num1=sc.nextFloat();
 		
 		System.out.print("Introduce el segundo número: ");
 		
-		num2=sc.nextInt();
+		num2=sc.nextFloat();
 		
 		System.out.println("\nOperaciones aritméticas básicas:" +
 						   "\nSuma: " + (num1+num2) +
 						   "\nResta: " + (num1-num2) +
 						   "\nMultiplicación: " + (num1*num2) +
 						   "\nDivisión: " + (num1/num2));
+		
 		sc.close();
 	}
